Add files via upload
I have used the font Montserrat and since georgia is not available on canva I had to use georgia pro. If any changes is to be done please inform. I am not able to  upload canva ppt through link so have converted to microsoft powerpoint version. Any recommendation or changes is highly aprreciated